// src/extension_core/wcp/generated/wcp_types.ts
export type DisplayedItemRef = string | number;
export type ItemPath = string;

export interface MarkerInfo {
  time: number;
  name?: string;
  move_focus: boolean;
}

export interface ItemInfo {
  name: string;
  type: string;
  id: DisplayedItemRef;
}

export interface WcpGreeting {
  type: "greeting";
  version: string;
  commands: Array<string>;
}

export interface WcpError {
  type: "error";
  message?: string;
}

export interface GetItemListCommand {
  type: "command";
  command: "get_item_list";
}

export interface GetItemInfoParams {
  ids: Array<DisplayedItemRef>;
}
// ...
export interface GetItemListResponse {
  type: "response";
  command: "get_item_list";
  ids: Array<DisplayedItemRef>;
}

export interface GetItemInfoResponse {
  type: "response";
  command: "get_item_info";
  results: Array<ItemInfo>;
}